requirements.txt
rembg
pillow
fastapi
onnxruntime
torch
transformers
trimesh 
pyrender 
pyglet 
numpy 
Pillow
pydantic-settings
uvicorn
python-multipart
peft
xformers
torchcodec